Backend/requirements.txt
fastapi>=0.116.0,<1.0.0
uvicorn[standard]>=0.35.0,<1.0.0
sqlalchemy>=2.0.41,<3.0.0
psycopg[binary]>=3.2.9,<4.0.0
pydantic-settings>=2.10.1,<3.0.0
python-jose[cryptography]>=3.5.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt==4.0.1
python-multipart>=0.0.20,<1.0.0
email-validator>=2.2.0,<3.0.0
openpyxl>=3.1.5,<4.0.0
python-docx>=1.1.2,<2.0.0
